refactor(streaming): extract usePointerLock hook
The pointer-lock request gate, the blocked-seq counter, and the cooldown
guard move out of the provider into their own hook. Takes the container
ref + the pause state + the connection-lost flag (the two things the
gate consults) and returns `{ blockedSeq, request, exit }`.

The cooldown logic — refusing the request when paused-but-not-yet
unlockable, and incrementing `blockedSeq` so consumers can play feedback
sounds — is now scoped to the hook rather than spread between
`requestPointerLock` and `setPointerLockBlockedSeq` at provider level.
